vim: less distracting and more informative folds
Folds now start with the amount of spaces that reflects fold level.
Number of folded lines was moved to the end.
With this it's pretty easy to distinguish nested folds.

Another improvement is that fold line will ignore
everything from the beginning of the line until after fold marker.
"// {{{6 Some comment" will be shown as "            Some comment"
(and the amount of folded lines after that)