vim: replace vim-plug with paq
Link: https://github.com/savq/paq-nvim

Main reason for switching is commands. With vim-plug I had separate
commands for updating plugins and updating plugin manager itself; and
there was no "all in one" command, that would install, update and clean
at the same time. I guess it's not hard to implement, but paq has it by
default.

And it's a step in migrating to lua-based configuration. Not sure if
I want to go all the way, but some pieces will be in lua anyway.

The easiest way to switch smoothly is to first add paq to vim-plug
and install it, then apply this commit, install everything with paq
and then clean vim-plug.