vim: add fzf configurations
Add a mapping for "command palette". This should've been ctrl+shift+p,
but either vim or alacritty can't distinguish it from ctrl+p.
So this is F7 just because it's easily accessible on my keyboard.
And move UndoTree to F12 as I don't use it that often.

Also an experimental feature for me: open vim's fzf window
in a tmux popup. This means that it'll have different colors
(and transparency).