vim: Add my own keymap wrapper
Most keymaps I set only need two things: a mapping and a command, so I
don't want to set mode every time. And I also want some other default
options like `noremap`.

This function has a third mandatory parameter: description. However
clean the command looks, I want to have a clear human test describing
what this does. One of the reasons is that if the command is a lua
function then it won't show up normally in a list of mappings, it'll be
a link to where this function is defined. But even simpler stuff is
easier to read then to parse.

Lsp was the primary place where I defined custom mappings, so it has the
most changes. I even moved the mappings for lsp to another file.
Along the way I added some more changes there and now
don't want to spend time on moving them to a dedicated commit:
* Show lightbulb! If current line has an action suggestion from the
  language server, a lightbulb appears in the sign column. It has lower
  priority than errors, I checked;
* Switched to real upstream for lsp_lines;
* Removed the config for `lsp_lines`. This plugin doesn't require
  calling `setup` function anymore;

Other notable changes:
* Changed commands in most lsp mappings to [fzf](https://github.com/ibhagwan/fzf-lua) (from vim built-ins or
  from lspsaga);
* "Go to definition" works through fzf too. It uses built in
  `--select-1` flag to have the same cool "open the definition if you're
  sure and open a list if not". There's one catch though: if there's
  only one definition, fzf still blinks, you can see that there was
  something between key press and the result;
* The fzf mappings from above have previews!
* New key mapping: toggle inlay hints. They are quite useful, but
  annoying, so I don't want to see them all the time;