vim: Extract some basic configs from init.vim
Some autocmds can be just filetype-specific configs, and some basic
keymaps can be in one place too.
This is also next step (finally) in the direction of init.vim ->
init.lua migration. Not that it's really necessary, but more convenient.

Turned out you can list all the modes you need for a keymap! So The type
of `mode` parameter is updated accordingly